+++ lib/alf/operator/relational/infer_heading.rb
@@ -1,13 +1,13 @@
 module Alf
   module Operator
     module Relational
-      class Heading
+      class InferHeading
         include Operator, Relational, Unary, Experimental
 
         signature do |s|
         end
 
-      end # class Project
+      end # class InferHeading
     end # module Relational
   end # module Operator
 end # module Alf