+++ spec/unit/alf-shell/operator/test_infer_heading.rb
@@ -1,14 +1,14 @@
 require 'spec_helper'
 module Alf::Shell::Operator
-  describe Heading do
+  describe InferHeading do
 
     let(:input){ [] }
-    subject{ Heading.run(argv) }
+    subject{ InferHeading.run(argv) }
 
     context "the default config" do
       let(:argv){ [input] }
       specify{
-        subject.should be_a(Alf::Operator::Relational::Heading)
+        subject.should be_a(Alf::Operator::Relational::InferHeading)
       }
     end